[Improvement]: getFieldDefinition should return localizedfields too
