+++ models/DataObject/ClassDefinition/Data/Localizedfields.php
@@ -28,10 +28,11 @@ use Pimcore\Model\Element;
 use Pimcore\Normalizer\NormalizerInterface;
 use Pimcore\Tool;
 
-class Localizedfields extends Data implements CustomResourcePersistingInterface, TypeDeclarationSupportInterface, NormalizerInterface, DataContainerAwareInterface, IdRewriterInterface, PreGetDataInterface, VarExporterInterface
+class Localizedfields extends Data implements CustomResourcePersistingInterface, TypeDeclarationSupportInterface, NormalizerInterface, DataContainerAwareInterface, IdRewriterInterface, PreGetDataInterface, VarExporterInterface, FieldDefinitionEnrichmentModelInterface
 {
     use Layout\Traits\LabelTrait;
     use DataObject\Traits\ClassSavedTrait;
+    use DataObject\Traits\FieldDefinitionEnrichmentModelTrait;
 
     /**
      * Static type of this element
@@ -96,11 +97,6 @@ class Localizedfields extends Data implements CustomResourcePersistingInterface,
      * @internal
      */
     protected array $referencedFields = [];
-
-    /**
-     * @internal
-     */
-    public ?array $fieldDefinitionsCache = null;
 
     /**
      * @internal
@@ -388,7 +384,7 @@ class Localizedfields extends Data implements CustomResourcePersistingInterface,
     public function setChildren(array $children): static
     {
         $this->children = $children;
-        $this->fieldDefinitionsCache = null;
+        $this->setFieldDefinitions([]);
 
         return $this;
     }
@@ -407,7 +403,7 @@ class Localizedfields extends Data implements CustomResourcePersistingInterface,
     public function addChild(mixed $child): void
     {
         $this->children[] = $child;
-        $this->fieldDefinitionsCache = null;
+        $this->setFieldDefinitions([]);
     }
 
     /**
@@ -416,7 +412,7 @@ class Localizedfields extends Data implements CustomResourcePersistingInterface,
     public function setReferencedFields(array $referencedFields): void
     {
         $this->referencedFields = $referencedFields;
-        $this->fieldDefinitionsCache = null;
+        $this->setFieldDefinitions([]);
     }
 
     /**
@@ -430,7 +426,7 @@ class Localizedfields extends Data implements CustomResourcePersistingInterface,
     public function addReferencedField(Data $field): void
     {
         $this->referencedFields[] = $field;
-        $this->fieldDefinitionsCache = null;
+        $this->setFieldDefinitions([]);
     }
 
     public function save(Localizedfield|AbstractData|\Pimcore\Model\DataObject\Objectbrick\Data\AbstractData|Concrete $object, array $params = []): void
@@ -589,60 +585,9 @@ class Localizedfields extends Data implements CustomResourcePersistingInterface,
     }
 
     /**
-     * @param string $name
-     * @param array $context additional contextual data
-     *
-     * @return DataObject\ClassDefinition\Data|null
+     * @internal
      */
-    public function getFieldDefinition(string $name, array $context = []): ?Data
-    {
-        $fds = $this->getFieldDefinitions($context);
-        if (isset($fds[$name])) {
-            $fieldDefinition = $fds[$name];
-            if (!\Pimcore::inAdmin() || (isset($context['suppressEnrichment']) && $context['suppressEnrichment'])) {
-                return $fieldDefinition;
-            }
-
-            $fieldDefinition = $this->doEnrichFieldDefinition($fieldDefinition, $context);
-
-            return $fieldDefinition;
-        }
-
-        return null;
-    }
-
-    /**
-     * @param array $context additional contextual data
-     *
-     * @return Data[]
-     */
-    public function getFieldDefinitions(array $context = []): array
-    {
-        if (empty($this->fieldDefinitionsCache)) {
-            $definitions = $this->doGetFieldDefinitions();
-            foreach ($this->getReferencedFields() as $rf) {
-                if ($rf instanceof DataObject\ClassDefinition\Data\Localizedfields) {
-                    $definitions = array_merge($definitions, $this->doGetFieldDefinitions($rf->getChildren()));
-                }
-            }
-
-            $this->fieldDefinitionsCache = $definitions;
-        }
-
-        if (!\Pimcore::inAdmin() || (isset($context['suppressEnrichment']) && $context['suppressEnrichment'])) {
-            return $this->fieldDefinitionsCache;
-        }
-
-        $enrichedFieldDefinitions = [];
-        foreach ($this->fieldDefinitionsCache as $key => $fieldDefinition) {
-            $fieldDefinition = $this->doEnrichFieldDefinition($fieldDefinition, $context);
-            $enrichedFieldDefinitions[$key] = $fieldDefinition;
-        }
-
-        return $enrichedFieldDefinitions;
-    }
-
-    private function doEnrichFieldDefinition(Data $fieldDefinition, array $context = []): Data
+    public function doEnrichFieldDefinition(Data $fieldDefinition, array $context = []): Data
     {
         if ($fieldDefinition instanceof FieldDefinitionEnrichmentInterface) {
             $context['class'] = $this;
@@ -650,33 +595,6 @@ class Localizedfields extends Data implements CustomResourcePersistingInterface,
         }
 
         return $fieldDefinition;
-    }
-
-    private function doGetFieldDefinitions(mixed $def = null, array $fields = []): array
-    {
-        if ($def === null) {
-            $def = $this->getChildren();
-        }
-
-        if (is_array($def)) {
-            foreach ($def as $child) {
-                $fields = array_merge($fields, $this->doGetFieldDefinitions($child, $fields));
-            }
-        }
-
-        if ($def instanceof DataObject\ClassDefinition\Layout) {
-            if ($def->hasChildren()) {
-                foreach ($def->getChildren() as $child) {
-                    $fields = array_merge($fields, $this->doGetFieldDefinitions($child, $fields));
-                }
-            }
-        }
-
-        if ($def instanceof DataObject\ClassDefinition\Data) {
-            $fields[$def->getName()] = $def;
-        }
-
-        return $fields;
     }
 
     public function getCacheTags(mixed $data, array $tags = []): array
